+++ test/tornado_tests/test_motor_transaction.py
@@ -48,7 +48,7 @@ from tornado.testing import gen_test
 
 from test import SkipTest
 from test.test_environment import env
-from test.tornado_tests import MotorTest
+from test.tornado_tests import AsyncVersion, MotorTest
 
 # Location of JSON test specifications.
 _TEST_PATH = os.path.join(
@@ -309,80 +309,6 @@ def end_sessions(sessions):
         s.end_session()
 
 
-class Version(tuple):
-    """Copied from PyMongo's test.version submodule."""
-
-    def __new__(cls, *version):
-        padded_version = cls._padded(version, 4)
-        return super(Version, cls).__new__(cls, tuple(padded_version))
-
-    @classmethod
-    def _padded(cls, iter, length, padding=0):
-        l = list(iter)
-        if len(l) < length:
-            for _ in range(length - len(l)):
-                l.append(padding)
-        return l
-
-    @classmethod
-    def from_string(cls, version_string):
-        mod = 0
-        bump_patch_level = False
-        if version_string.endswith("+"):
-            version_string = version_string[0:-1]
-            mod = 1
-        elif version_string.endswith("-pre-"):
-            version_string = version_string[0:-5]
-            mod = -1
-        elif version_string.endswith("-"):
-            version_string = version_string[0:-1]
-            mod = -1
-        # Deal with '-rcX' substrings
-        if '-rc' in version_string:
-            version_string = version_string[0:version_string.find('-rc')]
-            mod = -1
-        # Deal with git describe generated substrings
-        elif '-' in version_string:
-            version_string = version_string[0:version_string.find('-')]
-            mod = -1
-            bump_patch_level = True
-
-        version = [int(part) for part in version_string.split(".")]
-        version = cls._padded(version, 3)
-        # Make from_string and from_version_array agree. For example:
-        # MongoDB Enterprise > db.runCommand('buildInfo').versionArray
-        # [ 3, 2, 1, -100 ]
-        # MongoDB Enterprise > db.runCommand('buildInfo').version
-        # 3.2.0-97-g1ef94fe
-        if bump_patch_level:
-            version[-1] += 1
-        version.append(mod)
-
-        return Version(*version)
-
-    @classmethod
-    def from_version_array(cls, version_array):
-        version = list(version_array)
-        if version[-1] < 0:
-            version[-1] = -1
-        version = cls._padded(version, 3)
-        return Version(*version)
-
-    @classmethod
-    @gen.coroutine
-    def from_client(cls, client):
-        info = yield client.server_info()
-        if 'versionArray' in info:
-            raise gen.Return(cls.from_version_array(info['versionArray']))
-        raise gen.Return(cls.from_string(info['version']))
-
-    def at_least(self, *other_version):
-        return self >= Version(*other_version)
-
-    def __str__(self):
-        return ".".join(map(str, self))
-
-
 @gen.coroutine
 def should_run_on(scenario_def, client):
     run_on = scenario_def.get('runOn', [])
@@ -412,7 +338,7 @@ def should_run_on(scenario_def, client):
         if not version:
             raise gen.Return(True)
         version_tuple = tuple(int(elt) for elt in version.split('.'))
-        actual_version = yield Version.from_client(client)
+        actual_version = yield AsyncVersion.from_client(client)
         raise gen.Return(actual_version >= version_tuple)
 
     @gen.coroutine
@@ -421,7 +347,7 @@ def should_run_on(scenario_def, client):
         if not version:
             raise gen.Return(True)
         version_tuple = tuple(int(elt) for elt in version.split('.'))
-        actual_version = yield Version.from_client(client)
+        actual_version = yield AsyncVersion.from_client(client)
         raise gen.Return(actual_version <= version_tuple)
 
     for req in run_on:
